built a first Kite tile class
There is still a lot to work to do:

-Probably need to write a Vector class with all necessary operations to deal with addition, multiplication and maybe rotations.
- need to write __init__ methods for various situations

once that is done, I will need to create 'aware' tiles, that know not to creat new edges, but share existing edges.
To do that I will need to create a Tiling class, with edges, vertices, and tiles, which keeps track on what is on the frontier, and with grow methods.

Exciting!