Great Success! Tiles work
So the tiles are completely working.
Now I need to create an aware kind of tile, which knows about pre-existing edges and vertices, so it does not add new vertices, but shares them.
Or i could create a sort of 'builder class' that knows about existing vertices, edges and tiles so that it prevents the creation of useless edges and vertices, and prevents tiles from intersecting.