Litle change and added logic
Changed the assign method for the tiles, so they can be upgraded more easily. Added a Tile.definition attribute as well as Tile.assign_definition method, to be able to store the tile with less memory in the future.

Also, wrote a little piece of code that tells if extending a horizontal line to the right of a vertex will cross a given edge. This will be needed to see if a vertex lies within a tile.